chore: resolve circular dependency
